refactor: models and services to shared
