+++ apps/shared-models/badge.model.ts
@@ -1,15 +1,3 @@
-import { IAttachedFile } from './attached-file.model';
+import { IBadge } from '@commudle/shared-models';
 
-enum EBadgeTypes {
-  EXPERT = 'expert',
-  AUTHOR = 'author',
-}
-
-export interface IBadge {
-  // keeping id to any because it can be null
-  id: number;
-  name: string;
-  image: IAttachedFile;
-  slug: string;
-  badge_type: EBadgeTypes;
-}
+export { IBadge };
