+++ apps/shared-models/user_roles_user.model.ts
@@ -1,23 +1,3 @@
-import { ICommunity } from './community.model';
-import { IUser } from './user.model';
-import { IUserRole } from './user_role.model';
+import { EUserRolesUserStatus, IUserRolesUser } from '@commudle/shared-models';
 
-export interface IUserRolesUser {
-  id: number;
-  user_role: IUserRole;
-  user: IUser;
-  active: boolean;
-  parent_id: number;
-  parent_type: string;
-  parent_name: string;
-  status: EUserRolesUserStatus;
-  community?: ICommunity;
-  role_designation?: string;
-}
-
-export enum EUserRolesUserStatus {
-  INVITED = 'invited',
-  ACCEPTED = 'accepted',
-  REMOVED = 'removed',
-  JOINED_BY_TOKEN = 'joined_by_token',
-}
+export { IUserRolesUser, EUserRolesUserStatus };
