+++ libs/shared/models/package.json
@@ -1,0 +1,4 @@
+{
+  "name": "@commudle/shared-models",
+  "version": "0.0.1"
+}
