+++ libs/shared/models/src/index.ts
@@ -1,0 +1,12 @@
+export * from './lib/attached-file.model';
+export * from './lib/badge.model';
+export * from './lib/community.model';
+export * from './lib/community-build.model';
+export * from './lib/event.model';
+export * from './lib/event-status.model';
+export * from './lib/lab.model';
+export * from './lib/lab-step.model';
+export * from './lib/tag.model';
+export * from './lib/user.model';
+export * from './lib/user-role.model';
+export * from './lib/user-roles-user.model';
