+++ libs/shared/models/src/lib/badge.model.ts
@@ -1,0 +1,14 @@
+import { IAttachedFile } from './attached-file.model';
+
+enum EBadgeTypes {
+  EXPERT = 'expert',
+  AUTHOR = 'author',
+}
+
+export interface IBadge {
+  id: number;
+  name: string;
+  image: IAttachedFile;
+  slug: string;
+  badge_type: EBadgeTypes;
+}
