+++ libs/shared/models/src/lib/event-status.model.ts
@@ -1,0 +1,5 @@
+export interface IEventStatus {
+  id: number;
+  name: string;
+  description: string;
+}
