+++ libs/shared/models/src/lib/lab-step.model.ts
@@ -1,0 +1,5 @@
+export interface ILabStep {
+  id: number;
+  name: string;
+  description: string;
+}
