+++ libs/shared/models/src/lib/tag.model.ts
@@ -1,0 +1,10 @@
+import { IAttachedFile } from './attached-file.model';
+
+export interface ITag {
+  id: number;
+  name: string;
+  object_count?: number;
+  category?: string;
+  color?: string;
+  picture?: IAttachedFile;
+}
