+++ libs/shared/models/src/lib/user-role.model.ts
@@ -1,0 +1,4 @@
+export interface IUserRole {
+  id: number;
+  name: string;
+}
