+++ libs/shared/services/README.md
@@ -1,0 +1,7 @@
+# shared-services
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test shared-services` to execute the unit tests.
