+++ libs/shared/services/ng-package.json
@@ -1,0 +1,7 @@
+{
+  "$schema": "../../../node_modules/ng-packagr/ng-package.schema.json",
+  "dest": "../../../dist/libs/shared/services",
+  "lib": {
+    "entryFile": "src/index.ts"
+  }
+}
