+++ libs/shared/services/package.json
@@ -1,0 +1,11 @@
+{
+  "name": "@commudle/shared-services",
+  "version": "0.0.1",
+  "peerDependencies": {
+    "@angular/common": "^14.2.0",
+    "@angular/core": "^14.2.0"
+  },
+  "dependencies": {
+    "tslib": "^2.3.0"
+  }
+}
