+++ libs/shared/services/src/index.ts
@@ -1,0 +1,2 @@
+export * from './lib/api-routes.constant';
+export * from './lib/base-api.service';
