+++ tsconfig.base.json
@@ -16,7 +16,9 @@
     "baseUrl": ".",
     "paths": {
       "@commudle/auth": ["libs/auth/src/index.ts"],
-      "@commudle/ngx-datatable": ["libs/ngx-datatable/src/index.ts"]
+      "@commudle/ngx-datatable": ["libs/ngx-datatable/src/index.ts"],
+      "@commudle/shared-models": ["libs/shared/models/src/index.ts"],
+      "@commudle/shared-services": ["libs/shared/services/src/index.ts"]
     }
   },
   "exclude": ["node_modules", "tmp"]
